Extract create_tab1 into dedicated module
